fix unit tests

+++ lab_07_4/inc/sort_func.h
@@ -5,6 +5,6 @@ typedef int (*compar_fn_t)(const void *, const void *);
 
 int int_comparator(const void *a, const void *b);
 
-int mysort(void *data_array, size_t num, size_t size, compar_fn_t comparator);
+int mysort(void *data_array, int num, int size, compar_fn_t comparator);
 
 #endif // SORT_FUNC_H
